only makes sense on localhost

+++ test/nginx_system_test.sh
@@ -206,31 +206,33 @@ check [ $NUM_404_FINAL -eq $NUM_404_REALLY_FINAL ]
 
 start_test Non-local access to statistics fails.
 
-NON_LOCAL_IP=$(ifconfig | egrep -o 'inet addr:[0-9]+.[0-9]+.[0-9]+.[0-9]+' \
-                 | awk -F: '{print $2}' | grep -v ^127 | head -n 1)
+# This test only makes sense if you're running tests against localhost.
+if [ "$HOSTNAME" = "localhost:$PRIMARY_PORT" ] ; then
+  NON_LOCAL_IP=$(ifconfig | egrep -o 'inet addr:[0-9]+.[0-9]+.[0-9]+.[0-9]+' \
+    | awk -F: '{print $2}' | grep -v ^127 | head -n 1)
 
-# Make sure pagespeed is listening on NON_LOCAL_IP.
-URL="http://$NON_LOCAL_IP:$PRIMARY_PORT/mod_pagespeed_example/styles/"
-URL+="W.rewrite_css_images.css.pagespeed.cf.Hash.css"
-OUT=$(wget -O - -q $URL)
-check_from "$OUT" grep background-image
+  # Make sure pagespeed is listening on NON_LOCAL_IP.
+  URL="http://$NON_LOCAL_IP:$PRIMARY_PORT/mod_pagespeed_example/styles/"
+  URL+="W.rewrite_css_images.css.pagespeed.cf.Hash.css"
+  OUT=$(wget -O - -q $URL)
+  check_from "$OUT" grep background-image
 
+  # Make sure we can't load statistics from NON_LOCAL_IP.
+  ALT_STAT_URL=$(echo $STATISTICS_URL | sed s#localhost#$NON_LOCAL_IP#)
 
-# Make sure we can't load statistics from NON_LOCAL_IP.
-ALT_STAT_URL=$(echo $STATISTICS_URL | sed s#localhost#$NON_LOCAL_IP#)
+  echo "wget $ALT_STAT_URL >& $TEMPDIR/alt_stat_url.$$"
+  wget $ALT_STAT_URL >& "$TEMPDIR/alt_stat_url.$$"
+  check [ $? = 8 ]
+  rm -f "$TEMPDIR/alt_stat_url.$$"
 
-echo "wget $ALT_STAT_URL >& $TEMPDIR/alt_stat_url.$$"
-wget $ALT_STAT_URL >& "$TEMPDIR/alt_stat_url.$$"
-check [ $? = 8 ]
-rm -f "$TEMPDIR/alt_stat_url.$$"
+  ALT_CE_URL="$ALT_STAT_URL.pagespeed.ce.8CfGBvwDhH.css"
+  wget -O - $ALT_CE_URL  >& "$TEMPDIR/alt_ce_url.$$"
+  check [ $? = 8 ]
 
-ALT_CE_URL="$ALT_STAT_URL.pagespeed.ce.8CfGBvwDhH.css"
-wget -O - $ALT_CE_URL  >& "$TEMPDIR/alt_ce_url.$$"
-check [ $? = 8 ]
-
-wget -O - --header="Host: $HOSTNAME" $ALT_CE_URL >& "$TEMPDIR/alt_ce_url.$$"
-check [ $? = 8 ]
-rm -f "$TEMPDIR/alt_ce_url.$$"
+  wget -O - --header="Host: $HOSTNAME" $ALT_CE_URL >& "$TEMPDIR/alt_ce_url.$$"
+  check [ $? = 8 ]
+  rm -f "$TEMPDIR/alt_ce_url.$$"
+fi
 
 start_test Accept bad query params and headers
